+++ apps/backend/src/libs/modules/socket/base-socket.module.ts
@@ -2,16 +2,19 @@ import { type Server as HttpServer } from "node:http";
 import { type Socket, Server as SocketServer } from "socket.io";
 
 import { HTTPMethod } from "~/libs/modules/http/http.js";
+import { fileService } from "~/modules/files/files.js";
 import { meetingService } from "~/modules/meetings/meetings.js";
 
 import { type Logger } from "../logger/logger.js";
 import {
 	AllowedOrigin,
+	ContentType,
 	SocketEvent,
 	SocketMessage,
 	SocketNamespace,
 } from "./libs/enums/enums.js";
 import {
+	type MeetingAudioSaveDto,
 	type MeetingSummaryActionItemsResponseDto,
 	type MeetingTranscriptionRequestDto,
 	type ServerToClientEvents,
@@ -36,6 +39,26 @@ class BaseSocketService implements SocketService {
 
 	private handleBotsConnection(socket: Socket): void {
 		this.logger.info(`BOT ${SocketMessage.CLIENT_CONNECTED} ${socket.id}`);
+
+		socket.on(SocketEvent.AUDIO_SAVE, async (payload: MeetingAudioSaveDto) => {
+			try {
+				this.logger.info(SocketMessage.AUDIO_SAVE_RECEIVED);
+
+				const createdFile = await fileService.create({
+					contentType: ContentType.AUDIO,
+					key: payload.key,
+					url: payload.url,
+				});
+
+				await meetingService.attachAudioFile(payload.meetingId, {
+					fileId: createdFile.id,
+				});
+
+				this.logger.info(SocketMessage.AUDIO_SAVE_RECEIVED);
+			} catch (error) {
+				this.logger.error(`${SocketMessage.AUDIO_SAVE_ERROR} ${String(error)}`);
+			}
+		});
 
 		socket.on(SocketEvent.JOIN_ROOM, async (meetingId: string) => {
 			this.logger.info(`Client ${socket.id} joining room: ${meetingId}`);
@@ -92,7 +115,6 @@ class BaseSocketService implements SocketService {
 			async (payload: MeetingTranscriptionRequestDto) => {
 				try {
 					this.logger.info(SocketMessage.SOCKET_EVENT_RECEIVED);
-
 					const transcription = await meetingService.saveChunk(payload);
 
 					if (payload.meetingId) {