+++ apps/bot/src/libs/modules/socket-client/libs/types/client-to-server-events.type.ts
@@ -1,10 +1,12 @@
 import {
+	type MeetingAudioSaveDto,
 	type MeetingSummaryActionItemsResponseDto,
 	type MeetingTranscriptionRequestDto,
 } from "~/libs/types/types.js";
 
 type ClientToServerEvents = {
 	getPublicUrl: (meetingId: string) => void;
+	audioSave: (data: MeetingAudioSaveDto) => void;
 	joinRoom: (meetingId: string) => void;
 	recordingStopped: (meetingId: string) => void;
 	saveSummaryActionItems: (data: MeetingSummaryActionItemsResponseDto) => void;